+++ src/Light/Geometry/Matrix.hs
@@ -35,7 +35,7 @@ infixl 9 !!, //
 infixl 6 |+|, |-|
 infixl 7 |*|, |*, |/, |*^, ^*|, |*!, !*|, |*., .*|
 
-data Matrix = Matrix (UArray Int Float)
+data Matrix = Matrix (UArray Int Double)
 
 instance Show Matrix where
   show = show . (^.elems)
@@ -43,36 +43,36 @@ instance Show Matrix where
 instance Read Matrix where
   readsPrec n s = let [(fs, r)] = readsPrec n s in [(matrix fs, r)]
 
-matrix :: [Float] -> Matrix
+matrix :: [Double] -> Matrix
 matrix = Matrix . array (0, 15) . zip [0..]
 
-(!!) :: Matrix -> (Int, Int) -> Float
+(!!) :: Matrix -> (Int, Int) -> Double
 (Matrix m) !! (r, c) = m ! (r*4 + c)
 
-(//) :: Matrix -> [((Int, Int), Float)] -> Matrix
+(//) :: Matrix -> [((Int, Int), Double)] -> Matrix
 (Matrix m) // u = Matrix $ m A.// map (\((r, c), f) -> (r*4+c, f)) u
 
-elems :: Lens' Matrix [Float]
+elems :: Lens' Matrix [Double]
 elems = lens (\m -> m^.rows.traversed)
              (\m e -> rows.traversed .~ e $ m)
 
-ix :: (Int, Int) -> Lens' Matrix Float
+ix :: (Int, Int) -> Lens' Matrix Double
 ix p = lens (!! p)
             (\m v -> m // [(p, v)])
 
-row :: Int -> Lens' Matrix [Float]
+row :: Int -> Lens' Matrix [Double]
 row r = lens (\m -> [m !! (r, c) | c <- [0..3]])
              (\m vs -> m // [((r, c), v) | c <- [0..3] | v <- vs])
 
-col :: Int -> Lens' Matrix [Float]
+col :: Int -> Lens' Matrix [Double]
 col c = lens (\m -> [m !! (r, c) | r <- [0..3]])
              (\m vs -> m // [((r, c), v) | r <- [0..3] | v <- vs])
 
-rows :: Lens' Matrix [[Float]]
+rows :: Lens' Matrix [[Double]]
 rows = lens (\m -> [m^.row i | i <- [0..3]])
             (\_ vs -> matrix $ concat vs)
 
-cols :: Lens' Matrix [[Float]]
+cols :: Lens' Matrix [[Double]]
 cols = lens (\m -> [m^.col i | i <- [0..3]])
             (\_ vs -> matrix $ concat $ L.transpose vs)
 
@@ -94,7 +94,7 @@ identityMatrix = matrix [ 1, 0, 0, 0
 transpose :: Matrix -> Matrix
 transpose m = (cols .~ (m^.rows)) m
 
-dot :: [Float] -> [Float] -> Float
+dot :: [Double] -> [Double] -> Double
 dot x y = sum $ zipWith (*) x y
 
 (|+|), (|-|) :: Matrix -> Matrix -> Matrix
@@ -104,7 +104,7 @@ m |-| n = matrix $ zipWith (-) (m^.elems) (n^.elems)
 (|*|) :: Matrix -> Matrix -> Matrix
 m |*| n = matrix [ r `dot` c | r <- m^.rows, c <- n^.cols ]
 
-(|*), (|/) :: Matrix -> Float -> Matrix
+(|*), (|/) :: Matrix -> Double -> Matrix
 m |* s = (elems.traversed  *~ s) m
 m |/ s = (elems.traversed //~ s) m
 
@@ -138,25 +138,25 @@ translationMatrix v = matrix [ 1, 0, 0, v^.dx
                              , 0, 0, 1, v^.dz
                              , 0, 0, 0,     1 ]
 
-rotationMatrix :: Float -> Vector -> Matrix
+rotationMatrix :: Double -> Vector -> Matrix
 rotationMatrix angle axis = matrix [ u*u*(1-c)+c,   u*v*(1-c)-w*s, u*w*(1-c)+v*s, 0
                                    , v*u*(1-c)+w*s, v*v*(1-c)+c,   v*w*(1-c)-u*s, 0
                                    , w*u*(1-c)-v*s, w*v*(1-c)+u*s, w*w*(1-c)+c,   0
                                    , 0,             0,             0,             1 ]
   where c = cos angle; s = sin angle; u = axis^.dx; v = axis^.dy; w = axis^.dz
 
-frustumMatrix :: Float -> Float -> Float -> Float -> Float -> Float -> Matrix
+frustumMatrix :: Double -> Double -> Double -> Double -> Double -> Double -> Matrix
 frustumMatrix l r b t n f = matrix [ (2*n)/(r-l),           0,  a , 0
                                    ,           0, (2*n)/(t-b),  b', 0
                                    ,           0,           0,  c , d
                                    ,           0,           0, -1 , 0 ]
   where a = (r+l)/(r-l); b' = (t+b)/(t-b); c = -(f+n)/(f-n); d = -(2*f*n)/(f-n)
 
-perspectiveMatrix :: Float -> Float -> Float -> Float -> Matrix
+perspectiveMatrix :: Double -> Double -> Double -> Double -> Matrix
 perspectiveMatrix fovY aspect near far = frustumMatrix (-fW) fW (-fH) fH near far
   where fH = tan (fovY / 2) * near; fW = fH * aspect
 
-orthoMatrix :: Float -> Float -> Float -> Float -> Float -> Float -> Matrix
+orthoMatrix :: Double -> Double -> Double -> Double -> Double -> Double -> Matrix
 orthoMatrix l r b t n f = matrix [ 2/(r-l),       0,        0, tx
                                  ,       0, 2/(t-b),        0, ty
                                  ,       0,       0, -2*(f-n), tz
