+++ src/Light/Geometry/Normal.hs
@@ -14,14 +14,14 @@ import Control.Lens
 
 import Light.Geometry.Vector
 
-data Normal = Normal { _nx :: Float, _ny :: Float, _nz :: Float } deriving (Show, Read)
+data Normal = Normal { _nx :: Double, _ny :: Double, _nz :: Double } deriving (Show, Read)
 
 makeLenses ''Normal
 
-normal :: Float -> Float -> Float -> Normal
+normal :: Double -> Double -> Double -> Normal
 normal = Normal
 
-ns :: Lens' Normal [Float]
+ns :: Lens' Normal [Double]
 ns = lens (\ (Normal x y z) -> [x, y, z, 0]) (\_ [x, y, z, 0] -> Normal x y z)
 
 instance Eq Normal where
@@ -34,14 +34,14 @@ fromVector v = Normal (v^.dx) (v^.dy) (v^.dz)
 (Normal x y z) !+! (Normal a b c) = Normal (x+a) (y+b) (z+c)
 (Normal x y z) !-! (Normal a b c) = Normal (x-a) (y-b) (z-c)
 
-(!.!) :: Normal -> Normal -> Float
+(!.!) :: Normal -> Normal -> Double
 (Normal x y z) !.! (Normal a b c) = (x*a) + (y*b) + (z*c)
 
-(!*), (!/) :: Normal -> Float -> Normal
+(!*), (!/) :: Normal -> Double -> Normal
 (Normal x y z) !* s = Normal (x*s) (y*s) (z*s)
 (Normal x y z) !/ s = Normal (x/s) (y/s) (z/s)
 
-(*!) :: Float -> Normal -> Normal
+(*!) :: Double -> Normal -> Normal
 s *! (Normal x y z) = Normal (x*s) (y*s) (z*s)
 
 infixl 6 !+!, !-!
@@ -50,7 +50,7 @@ infixl 7 !.!, !*, !/, *!
 negateN :: Normal -> Normal
 negateN (Normal a b c) = Normal (-a) (-b) (-c)
 
-magnitudeN, magnitudeSquaredN :: Normal -> Float
+magnitudeN, magnitudeSquaredN :: Normal -> Double
 magnitudeN = sqrt . magnitudeSquaredN
 magnitudeSquaredN n = n !.! n
